refactor app module
